chore: move /public to rrot dir & update imports
