+++ src/components/UI/HomepageHero.tsx
@@ -5,8 +5,8 @@ import { FC } from 'react';
 import { PageSubheading } from '../UI';
 import { ButtonLink } from '../../components';
 
-import homepageHero from '../../public/images/homepage-hero.svg';
-import homepageHeroMobile from '../../public/images/homepage-hero-mobile.svg';
+import homepageHero from '../../../public/images/homepage-hero.svg';
+import homepageHeroMobile from '../../../public/images/homepage-hero-mobile.svg';
 
 import { APPLICANTS_URL } from '../../constants';
