+++ src/components/UI/Step/StepArrow.tsx
@@ -2,7 +2,7 @@ import { Box } from '@chakra-ui/react';
 import Image from 'next/image';
 import { FC } from 'react';
 
-import stepPolygonSVG from '../../../public/images/step-polygon.svg';
+import stepPolygonSVG from '../../../../public/images/step-polygon.svg';
 
 export const StepArrow: FC = () => {
   return (
