+++ src/components/forms/SmallGrantsForm.tsx
@@ -27,7 +27,7 @@ import { useShadowAnimation } from '../../hooks';
 
 import { chakraStyles } from './selectStyles';
 
-import planeVectorSVG from '../../public/images/plane-vector.svg';
+import planeVectorSVG from '../../../public/images/plane-vector.svg';
 
 import {
   COMMUNITY_EVENT,
